+++ script/c64496451.lua
@@ -160,10 +160,10 @@ function c64496451.spop(e,tp,eg,ep,ev,re,r,rp)
 		e1:SetType(EFFECT_TYPE_FIELD+EFFECT_TYPE_CONTINUOUS)
 		e1:SetCode(EVENT_PHASE+PHASE_END)
 		e1:SetProperty(EFFECT_FLAG_IGNORE_IMMUNE)
-		e1:SetReset(RESET_PHASE+PHASE_END)
 		e1:SetCountLimit(1)
 		e1:SetLabel(fid)
 		e1:SetLabelObject(sg)
+		e1:SetCondition(c64496451.descon)
 		e1:SetOperation(c64496451.desop)
 		Duel.RegisterEffect(e1,tp)
 	end
@@ -171,10 +171,17 @@ end
 function c64496451.desfilter(c,fid)
 	return c:GetFlagEffectLabel(64496451)==fid
 end
+function c64496451.descon(e,tp,eg,ep,ev,re,r,rp)
+	local g=e:GetLabelObject()
+	if not g:IsExists(c64496451.desfilter,1,nil,e:GetLabel()) then
+		g:DeleteGroup()
+		e:Reset()
+		return false
+	else return true end
+end
 function c64496451.desop(e,tp,eg,ep,ev,re,r,rp)
 	local g=e:GetLabelObject()
 	local tg=g:Filter(c64496451.desfilter,nil,e:GetLabel())
-	g:DeleteGroup()
 	Duel.Destroy(tg,REASON_EFFECT)
 end
 function c64496451.valcheck(e,c)
